Feat enable coupon

+++ lib/payment/service/payment_service.dart
@@ -71,7 +71,7 @@ class PaymentService {
     required BuildContext context,
     // required String description,
     // required int amount,
-    required int InvoiceValue,
+    required double InvoiceValue,
   }) async {
     final getStorage = GetStorage();
     String token = getStorage.read('accessToken') ?? "";
@@ -169,7 +169,7 @@ class PaymentService {
     required BuildContext context,
     required String language,
     required String paymentMethod,
-    required int price,
+    required double price,
   }) async {
     final getStorage = GetStorage();
     String token = getStorage.read('accessToken') ?? "";
@@ -212,7 +212,7 @@ class PaymentService {
 
   static Future<Invoice?> creditCardEmbedded({
     required BuildContext context,
-    required int price,
+    required double price,
   }) async {
     final getStorage = GetStorage();
     String token = getStorage.read('accessToken') ?? "";
@@ -328,7 +328,7 @@ class PaymentService {
   }
 
   static Future<Invoice?> applePayEmbedded(
-      {required BuildContext context, required int invoiceValue}) async {
+      {required BuildContext context, required double invoiceValue}) async {
     final getStorage = GetStorage();
     String token = getStorage.read('accessToken') ?? "";
     if (JwtDecoder.isExpired(token)) {
